Change Point2f to Point2i

+++ Visio/cpp/src/Blobs.cpp
@@ -88,7 +88,7 @@ void Blobs::Relier(){
 cv::Mat Blobs::Get_img_sep() const{return img_sep;}
 cv::Mat Blobs::Get_img_blobs() const{return img_blobs;}
 std::vector <cv::Moments> Blobs::Get_mu() const{return mu;}
-std::vector <cv::Point2f> Blobs::Get_mc() const{return mc;}
+std::vector <cv::Point2i> Blobs::Get_mc() const{return mc;}
 std::vector <cv::Rect> Blobs::Get_rect() const{return rect;}
 std::vector <double> Blobs::Get_size() const{return taille;}
 void Blobs::Set_img(cv::Mat image){
