+++ Visio/cpp/src/Gui.h
@@ -32,15 +32,15 @@ public:
 	STRUCT_HSV_BOUND *Get_HSV_bound() const;					// pointeur vers la structure contenant les paramètres de séparation
 	STRUCT_HSV_BOUND *Get_HSV_bound2() const;					// pointeur vers la structure contenant les paramètres de séparation 2
 	STRUCT_WRAP_BOUND *Get_wrap_bound() const;					// pointeur vers la structure contenant les paramètres de transformation
-	void Controler_souris(std::vector <cv::Point2f> mc, int width, int height);	// contrôler la souris en fonction de la position du 1er blob
-	void Cliquer(std::vector <cv::Point2f> mc);					// cliquer si le vecteur n'est pas vide
+	void Controler_souris(std::vector <cv::Point2i> mc, int width, int height);	// contrôler la souris en fonction de la position du 1er blob
+	void Cliquer(std::vector <cv::Point2i> mc);					// cliquer si le vecteur n'est pas vide
 	void Afficher_image(const std::string titre_fenetre, cv::Mat image);
 	void Pad(const std::string titre_fenetre, const float dx, const float dy, const float dx_max, const float dy_max);
 	void Ajouter_vecteurs(
 		const std::string titre_fenetre,	// titre de la fenêtre
 		cv::Mat image,				// image à traiter
-		std::vector<cv::Point2f> pts_prev,	// tableau des points initiaux
-		std::vector<cv::Point2f> pts_next);	// tableau des points finals
+		std::vector<cv::Point2i> pts_prev,	// tableau des points initiaux
+		std::vector<cv::Point2i> pts_next);	// tableau des points finals
 
 private:
