+++ Visio/cpp/src/mains/sous-marin/metadonnees.cpp
@@ -55,7 +55,7 @@ int main(int argc, char *argv[]){
 		reco.Detecter_quadrillage();
 		if(key == 's'){
 			mkdir(("./output/" + to_string(compteur)).c_str(), S_IRWXU | S_IRWXG | S_IRWXO);
-			vector < vector <cv::Point2f> > liste_quadrillage = reco.Get_quadrillage();
+			vector < vector <cv::Point2i> > liste_quadrillage = reco.Get_quadrillage();
 			for(int i = 0; i < liste_quadrillage.size(); i++){
 				// transformation
 				cv::Mat img_redressee;
